% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_cv.R
\name{compute_auc}
\alias{compute_auc}
\title{Compute the NEGATIVE value of the area under the ROC curve (AUC) for the binary logistic model}
\usage{
compute_auc(y, mu)
}
\arguments{
\item{y}{A response vector of size \code{n}. Usually used as the response in the held-out set.}

\item{mu}{A predicted linear predictor values. Usually used as the linear predictions for the held-out set over a path of tuning parameter values, i.e., \code{mu} is a matrix of \code{n} rows and the number of tuning parameters columns.}
}
\value{
The output is the negative value of AUC for each value of tuning parameters used to compute each column of mu. We return the negative value to be consistent with the cross-validation procedure which selects the tuning parameter that minimizes CV-estimate of test error.

Estimate of AUC could be inaccurate if \code{n} is small.
}
\description{
Compute the NEGATIVE value of the area under the ROC curve (AUC) for the binary logistic model
}
