% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_cv.R
\name{compute_deviance_ordinal}
\alias{compute_deviance_ordinal}
\title{Compute deviance of the ordinal logistic regression model}
\usage{
compute_deviance_ordinal(y, mu)
}
\arguments{
\item{y}{A response vector of size \code{n}. Usually used as the response in the held-out set.}

\item{mu}{A predicted linear predictor values in the form of an \code{n}-by-\code{[object$nlev-1]} matrix. Usually used as the linear predictions for the held-out set over a path of tuning parameter values.}
}
\value{
The output is deviance of the ordinal logistic regression model for each value of tuning parameters.
}
\description{
Compute deviance of the ordinal logistic regression model
}
