% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_cv.R
\name{compute_deviance_poisson}
\alias{compute_deviance_poisson}
\title{Compute deviance of the Poisson regression model}
\usage{
compute_deviance_poisson(y, mu)
}
\arguments{
\item{y}{A response vector of size \code{n}. Usually used as the response in the held-out set.}

\item{mu}{A predicted linear predictor values. Usually used as the linear predictions for the held-out set over a path of tuning parameter values, i.e., \code{mu} is a matrix of \code{n} rows and the number of tuning parameters columns.}
}
\value{
The output is deviance of the Poisson regression model for each value of tuning parameters used to compute each column of mu.
}
\description{
Compute deviance of the Poisson regression model
}
