% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_cv.R
\name{compute_error}
\alias{compute_error}
\title{Compute the error used in each fold of cross-validation}
\usage{
compute_error(y, mu, family, type.measure)
}
\arguments{
\item{y}{A response vector of size \code{n}. Usually used as the response in the held-out set.}

\item{mu}{A predicted linear predictor values. Usually used as the linear predictions for the held-out set over a path of tuning parameter values.}

\item{family}{Either a character string representing one of the built-in exponential families. Default is "gaussian".}

\item{type.measure}{type of validation error used for cross-validation. Currently only supporting 2 options. The default is \code{type.measure = "deviance"}, which uses squared-error for gaussian models, deviance for logistic, poisson and ordinal logistic regression. \code{type.measure = "auc"} is for two-class logistic regression only, and gives area under the ROC curve.}
}
\description{
Compute the error used in each fold of cross-validation
}
