% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_lambda}
\alias{get_lambda}
\title{A general function for calculating the lambda sequence for lasso problem}
\usage{
get_lambda(
  x,
  y,
  offset = NULL,
  intercept = FALSE,
  family = "gaussian",
  nlam = 100,
  lam_min_ratio = ifelse(nrow(x) < ncol(x), 0.01, 1e-04)
)
}
\description{
A general function for calculating the lambda sequence for lasso problem
}
