% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.sprinter.R
\name{predict.sprinter}
\alias{predict.sprinter}
\title{Calculate prediction from a \code{sprinter} object.}
\usage{
\method{predict}{sprinter}(object, newdata, ...)
}
\arguments{
\item{object}{a fitted \code{sprinter} object.}

\item{newdata}{a design matrix of all the \code{p} main effects of some new observations of which predictions are to be made.}

\item{...}{additional argument (not used here, only for S3 generic/method consistency)}
}
\value{
The prediction of \code{newdata} by the sprinter fit \code{object}. The prediction is on the scale of the linear predictors, i.e., \code{type = "link"} as in \code{predict.glm} or \code{predict.glmnet}. For example, for "gaussian" family it gives the fitted response values, and for "binomial" family the predictions are of log-odds (logit).

 The output is a list consisting of \code{length(object$lambda1)} elements. For the "gaussian", "poisson", and "binomial" families, each element is a matrix of dimensions \code{n}-by-\code{nrow(object$lambda3)}. For "ordinal" family, each element is a list with \code{length(object$lambda3)} components, each of which is an \code{n}-by-\code{[object$nlev-1]} matrix.
}
\description{
Calculate prediction from a \code{sprinter} object.
}
\examples{
n <- 100
p <- 200
x <- matrix(rnorm(n * p), n, p)
y <- x[, 1] + 2 * x[, 2] - 3 * x[, 1] * x[, 2] + rnorm(n)
mod <- sprinter(x = x, y = y)
fitted <- predict(mod, newdata = x)

}
