% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sis_lasso.R
\name{sis_lasso}
\alias{sis_lasso}
\title{Sure independence screening followed by lasso}
\usage{
sis_lasso(x, y, family = "gaussian", num_keep = NULL, ...)
}
\arguments{
\item{x}{An \code{n} by \code{p} design matrix of main effects. Each row is an observation of \code{p} main effects.}

\item{y}{A response vector of size \code{n}.}

\item{family}{Either a character string representing one of the built-in exponential families, including "gaussian", "poisson", and "binomial". Default is "gaussian".}

\item{num_keep}{Number of variables to keep in the screening phase}

\item{...}{other arguments to be passed to the \code{glmnet} calls, such as \code{alpha} or \code{penalty.factor}}
}
\value{
An object of S3 class "\code{sis_lasso}".
 \describe{
  \item{\code{n}}{The sample size.}
  \item{\code{p}}{The number of main effects.}
  \item{\code{a0}}{Intercept value.}
  \item{\code{compact}}{A compact representation of the selected variables. \code{compact} has three columns, with the first two columns representing the indices of a selected variable (main effects with first index = 0), and the last column representing the estimate of coefficients.}
 }
}
\description{
Sure independence screening followed by lasso
}
\examples{
set.seed(123)
n <- 100
p <- 200
# dense input
x <- matrix(rnorm(n * p), n, p)
y <- x[, 1] - 2 * x[, 2] + 3 * x[, 1] * x[, 3] - 4 * x[, 4] * x[, 5] + rnorm(n)
mod <- sis_lasso(x = x, y = y)

}
